refactor albumlists-artist
